requirements.txt
requests
opencv-python
python-dotenv
scikit-learn
progressbar2